add tests relocate utils.py, and change path in main
